Separate into files and add integration tests
